refactor: TodoProvider 파일 내부에 선언된 타입  파일 분리
