+++ web/src/type/index.ts
@@ -1,2 +1,3 @@
 export type { TextFieldData } from "./field";
 export type { Method, UseFetchParams, PostParams, GetParams } from "./useFetch";
+export type { TodoTexts, TodoItem, ContextType } from "./todoProvider";
